requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
faiss-cpu>=1.7.0
PyMuPDF>=1.23.0
pillow>=10.0.0
nltk>=3.8.0
pandas>=2.0.0
